refactor data addition functions into db.py
